src/api/myfirm/requirements.txt
# This Version of modules is from Python 3.8 
fastapi==0.110.1
pydentic==2.7.0
mysql-connector==2.2.9